Lawn-and-Garden-Supplies-Market.md
Lawn and Garden Supplies Market, Global Outlook and Forecast 2023-2029 The global Lawn and Garden Supplies market was valued at US$ 11150 million in 2022 and is projected to reach US$ 13690 million by 2029, at a CAGR of 3.0% during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes.